examples/hermit: initial work-in-progress
Current status is that we do get the Hermit loader and kernel
to start booting. The kernel gets stuck when trying to initialising
the GIC. This makes sense since Hermit is expecting GICv3 to be
available, and libvmm is emulating GICv2. We will need to do a couple
things here, including re-building seL4 to use GICv3 for QEMU. This
*should* work but just takes some configuration.