examples/hermit: much more experimentation and debugging
Current status is that Hermit enters its scheduler (which is great since
it means that the kernel is almost functional), however it tries to
send an SGI to itself which ends up causing an exception. This seems
reasonable as the guest should not be able to create interrupts without
going through the hypervisor. Right now, to my knowledge, seL4 does not
provide a way to create SGIs from user-space. This means that the VMM
can't really do anything in this situation I think. We could detect
the exception and fake an SGI, but I'm not sure whether 1. that would
even work and 2. whether that's a good approach.

To my understanding we are now dealing with a limitation of seL4 itself
but am not 100% sure. More investigation required.